Fix model visibility toggles and refresh
